backend/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-groq
PyPDF2
